Update patch set 5
Patch Set 5:

(1 comment)

So, the owner is only stored when the uniform's resolvedName ends up different than the original name? (What does that imply? I'm trying to figure out the logic that leads to us adding a mapping or not). Also, I see that the owner needs to be an FP, and it looks like most places that insert uniforms from FPs are inserting the owner (and GP/XP locations aren't). The spot I tagged in the GrGLSLColorSpaceXformHelper is most often used in the context of an FP - should we add a GrFragmentProcessor* arg to emitCode there?

Patch-set: 5
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>