Use round for morphology radius
The fiddle: https://fiddle.skia.org/c/1a0202dd230713064ad90b627c7de47d
highlighted an issue with the old conversion from float to int for the
mapped dilate/erode radii. Depending on the canvas matrix and the scale
extraction floating point error, the computed float raddi would be
very close to 1.0 (for a nominal radii = 1), but actually be 0.9999.

The floor would turn this into a 0 integral radius and then the image
filter implementation would determine that no operation was needed.
Switching to round fixes the issue present in the filter. This may
cause some GMs to change as it does change how the floating point radii
are mapped. Now, for fractional radii, we will jump up to the next
kernel size a little sooner.

The SVG spec appears ambiguous in terms of rounding behavior for these
scenarios. The other kernel-based image filters that I checked
(matrix, blur, drop-shadow) are all defined in device space, or use
a sigma to determine kernel size, so morphology filters are somewhat
unique. This means we don't need to worry about creating a
discrepency between this filter's kernel rounding behavior and other
filters' behaviors.

Bug: skia:10066
Change-Id: I6b3daa4c1242273ce924e22ba6de5eb25a7a96bc