Update patch set 1
Patch Set 1:

Ben found this discussion: https://github.com/w3c/csswg-drafts/issues/3619, so it is somewhat on the radar of the CSS spec that the rounding behavior needs to be specified. I'll do a little poking to see if Chromium attempts to pre-round before it gets to skia (in which case we have a bit more flexibility in what we do).

Patch-set: 1
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0