Update patch set 3
Patch Set 3:

The benefit of the new function is for immediate filtering of single color rather than needing it at raster time.

e.g. If we want to set the background color for rectangle, setting value filtered value directly helps to encapsulate in paint ops used in chrome. This is typical need of dark mode feature in chrome.

Patch-set: 3