Update patch set 4
Patch Set 4: Code-Review-1

OK... this is exactly the sort of thing I've been trying to explain in the Chromium bug that we really don't want.  Please don't land this.

Performance changes to this filter to improve the performance of the filter's existing control flow are fine, but performance changes to this filter to improve only single color transform speed are not welcome.

Patch-set: 4
Reviewer: Gerrit User 5045 <5045@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=-1