Update patch set 14
Patch Set 14: -Code-Review

If we're going to go with one array owning everything and then other arrays with pointers, can you add a debug validate method that we call before all staging buffer operations. The method will check that no buffer is in more than one of the lists, all buffers in the owning array exist in a list, and that there is no buffer in any list that isn't in the owning array.

Also I realized with this change the GrGpu owns objects. So we need implement the disconnect() function in GrGpu and have it call a virtual onDisconnect to the backends. Annoyingly we need the backends to do their own synchronization to know the GPU has finished all the work, but we have to free everything before the end of onDisconnect. So I think we're stuck with the backends having to cleanup all the lists. After waiting on all GPU work we should assert in each backend that only the available list has buffers in it. The base class disconnect() should call onDisconnect and then assert that all lists on the owning array are empty.

Patch-set: 14
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=0