Attempt to repro b/153410582
Bug: b/153410582

On Android, these images decode fine. But then if you call
Bitmap#createScaledBitmap, the resulting Bitmap is all black. The
reporter suspects it is related to the color profile.

This CL attempts to reproduce the calls used in Android, but it
does not reproduce the bug. This means either my code is missing a
step or the bug is outside of Skia on Anroid.

Change-Id: I90528526d41c4e5b5f92f597e613229a46fecc0d