Update patch set 1
Patch Set 1:

Yeah I am pretty sure positive its a driver issue. I've stared at the test that is creating the VkDevice (it is very simple) and I've convinced myself we are destroying the VkDevice correctly. Additionally, this is the last line of loader we have in the stack trace before missing frames, https://github.com/KhronosGroup/Vulkan-Loader/blob/sdk-1.1.130/loader/loader.c#L6629. That is where the loader is calling into the driver (so the allocation is from the driver). Finally the leaked memory is 128 bytes. I would imagine the memory allocated for a VkDevice would be significantly higher and we'd see a larger leak if the whole VkDevice wasn't destroyed and leaked.

Patch-set: 1
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>