Update patch set 2
Patch Set 2:

I'm ok with this I suppose, but it is unfortunate that it means we have to keep the freetype suppression (which we could remove with full stack traces, since it's actually a fontconfig thing but we don't currently see fontconfig in here due to the truncated stacks). I'm guessing we have this issue because debian built without frame pointers?

Patch-set: 2
Reviewer: Gerrit User 5140 <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>