Update patch set 15
Patch Set 15: Code-Review+1

> Patch Set 14:
> 
> > Patch Set 14:
> > 
> > > Patch Set 14:
> > > 
> > > > Patch Set 14: -Code-Review
> > > > 
> > > > If we're going to go with one array owning everything and then other arrays with pointers, can you add a debug validate method that we call before all staging buffer operations. The method will check that no buffer is in more than one of the lists, all buffers in the owning array exist in a list, and that there is no buffer in any list that isn't in the owning array.
> 
> First is done, second is guaranteed by SkTInternalList, third is done but O(N^2). 😞
> 
> TBH, I find all this kind of busywork that makes the code noisy. It's pretty easy to see by inspection that these are all true.

I mean each backend is going to be moving things around these buffers so I don't think it is that easy to quickly look that they all are doing the correct thing especially since each will have a different method to know when the GPU is done. I generally perfer the machine to tell me I'm doing everything correct. We could have just left each list owning the unique_ptrs and then all of this is just guaranteed, but by wanting a different array to be the owner this is how we make sure things are working as expected.

> 
> > > > Also I realized with this change the GrGpu owns objects. So we need implement the disconnect() function in GrGpu and have it call a virtual onDisconnect to the backends. Annoyingly we need the backends to do their own synchronization to know the GPU has finished all the work, but we have to free everything before the end of onDisconnect. So I think we're stuck with the backends having to cleanup all the lists. After waiting on all GPU work we should assert in each backend that only the available list has buffers in it. The base class disconnect() should call onDisconnect and then assert that all lists on the owning array are empty.
> 
> Implemented in GrDawnGpu::disconnect(). I *think* I've met the contract for kAbandon as best I can, although since wgpu::Buffer is refcounted by Dawn, there's no way to avoid calling the unref, which could make backend API calls.

Yeah you'll have to answer the dawn abandon question. For example, in Vulkan we are required to cleanly destroy every object even in a device lost situation. Thus our Vulkan backend doesn't really support abandon and just always cleans up all its internals. Whereas in GL the GLContext may die and in that case it is fine to just drop all the GL objects on the floor. So you'll need to answer that question about dawn.

> 
> > > disconnect() is already virtual, and overridden in GrVkGpu and GrGLGpu. Do you want me to just put the validation in the back disconnect(), or rework the system to have a non-virtual disconnect and a virtual onDisconnect?
> > 
> > Actually leave as with diconnect being virtual and add the asserts that everything is empty in the GrGpu destructor.
> 
> Done.

Patch-set: 15
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1