Update patch set 6
Patch Set 6:

Not used, yet.

Currently YUVA images are specified by having a plane/texture index and a SkColorChannel value for each of Y, U, V, A. This leads to all sorts of confusion when textures have both a format and a color type. We conflate color type channels with texture format channels which isn't always an identity mapping. It's also tricky to get the channel correct for single channel textures when our code/client code doesn't know which channels are present in the texture which has caused us to add some bug prone transformations

The idea I'm developing here is to replace the SkColorChannel in the YUVA specification with an index in rang 0-3. The index refers to the 0th, 1st, ... channel. Each legal mask combination has an implicit order (e.g. RG is 0:R, 1:G. Gray is 0:Gray, Alpha is 0:Alpha, ...). This can apply either to SkColorTypes (for pixmaps) or backend formats (for textures) but we will never use a color type with a backend texture used to make a YUVA image, internally or in the API, so there is no confusion about which it refers to. And for single channel textures, channel index 0 always refers to that single channel, whether it's red, luminance, or alpha.

That said, the use of SkYUVAIndices in Chrome code is widespread and it will take a while to get to the end goal. There are other API considerations to resolve before replacing SkYUVAIndex in callers, such as how should chroma subsampling, should we use a simple enum of planar configs, ... Even if the public API is more like a simple enum I think we want to internally represent the planar configurations as these plane/channel index pairs to remove ambiguity for single channel formats.

Patch-set: 6
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Gerrit User 5230 <5230@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Gerrit User 5065 <5065@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Gerrit User 5085 <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>