Update patch set 1
Patch Set 1:

(2 comments)

This is not really directly related to this CL, but the move from protected to private made me think of it, and wanted to see your thoughts.

In the back of mind I've been trying to think of the details of implementing it on vulkan and I'm trying to decide if Vulkan wants the busy list or if it wants to do its own thing. There is a chance I'd move the busy list into GrDawnGpu and the backends do more specialized control during that part of the loop. In that world I'd have to move the active list might back to protected or have some sort of protected getter.

Basically what I'm thinking is the vulkan backend would move StagingBuffers off the active list and store the pointers directly on the GrVkCommandBuffer that we're submitting that frame (the command buffer is what we track the completion of on the gpu). So then when each command buffer finishes it can easily moves its set of staging buffers back to the available list. There is no searching through busy.

Certainly it can be done with the current busy list as well, still trying to think through all the details. Regardless, things I'll worry about when I actually implement something.

Patch-set: 1
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>