Order instructions by tree, and register pressure.
Implement a simplification of the scheduler described in
"Register-Sensitive Selection, Duplication, and Sequencing of
Instructions". Break up the DAG into trees rooted at sinks or nodes
that have a fan-out > 1. Schedule each tree optimally. The instructions
for trees are kept together using a tree number, and the instructions
of a tree are schedule optimally loosely based on "The Generation of
Optimal Code for Arithmetic Expressions"

Change-Id: Icd1fe687529e3815371fc56b6ce41ccf7172fa57