Try out plane/channel index scheme in wacky yuv format GMs.
The idea is to avoid specifying RGBA channels for the components of a
YUVA format until a backend format or color type is known. Instead the
YUVA format is defined by channel indices and then channels present in
the bitmap/texture are indexed.

Right now this is just local to a GM but the thought is that longer
term it would make sense to use this scheme in place of internal uses
of SkYUVAIndex.

I believe this also fixes a bug in the GM related to the packing of
Y410.

Bug: skia:10078

Change-Id: Ic47ac84ab7a047d91ae72b45ccd0adc9f449e2be