Try out plane/channel index scheme in wacky yuv format GMs.
The idea is to avoid choosing RGBA channels for YUVA components until
texture backend formats/bitmap color types are known. Instead the YUVA
format is represented using channel indices which are applied to backend
formats/color types when we have actual planes.

This is local to the GM now, but the thinking is that this could replace
internal usage of SkYUVAIndex. Possibly also public use, but perhaps an
enum instead of a table is a better API.

Also fixes a bug in the packing of Y410.

Bug: skia:10078

Change-Id: Ic47ac84ab7a047d91ae72b45ccd0adc9f449e2be