Update patch set 2
Patch Set 2:

Yeah, this definitely won't work as-is. I tried several variations, attempting to get the spurious save triggers out of onSaveCamera, but couldn't get it to work. So I went down the path of trying to decompose the saveCamera logic at the non-virtual SkCanvas level. Hence, the hack in NWay so it only triggers the camera stack update on the children (as they've already seen everything else from the regular concat call).

But for picture playback, we'd need a similar backdoor so that the record from the captured onSaveCamera doesn't trigger a full saveCamera, just the camera-stack logic.

Patch-set: 2