simplify Ymm/Xmm params in SkVM Assembler
I've been looking at the assembler and thinking of exposing some more of
the generality of x86 instruction encoding, adding abstractions like
YmmOrLabel to reflect other regular features like ModRM, SIB, etc.

One of the things I remembered is that the difference between
instructions on xmm and on ymm registers is generally just one bit in
VEX, and thought, maybe an XmmOrYmm type that handles that concept is a
good first warmup abstraction to introduce.

But before I even do that, this CL warms up the warmup.

Here I've switched the argument type to Ymm for all instructions that,
while the technically encoded as operating on an Xmm register, we always
cast a Ymm to Xmm when used and they're totally unambiguous.  Take vmovd
for example...  you're only ever working with the bottom 4 bytes, so it
doesn't really matter if you think of it as a Ymm or Xmm register.

This leaves Xmm exposed in the API in only small ways:

   - distinguishing vmovups(dst, Xmm) from vmovups(dst, Ymm)
     (this is where XmmOrYmm might first shine)

   - byte and word inserts and extracts,
     which only function on the low 128 bits.

Switching the byte/word insert/extract operations over to also talk Ymm
would not be ambiguous, but I thought requiring an Xmm register might
serve as a good reminder that you can only insert or extract from those
128 bits, and that you have to do fancier things for the high bits.

Change-Id: I68ec8cb620cbc6203dd758aafcbb999ba0fcf78e