Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> FYI, I minimized this and reported at https://developercommunity.visualstudio.com/content/problem/986671/ice-with-const-initializer-list-in-constexpr-conte.html .

Thanks! BTW when I wrote the CL that introduced this issue I found that the compiler version on the bots produced the same issue when the instances were declared with constexpr so I removed that. Then Rob discovered on his machine that the latest compiler produced the error when the instances were just static const.

Patch-set: 3