Add submits after flushes in ProgramsTest.
This looks to be the culprit behind the failing/flaky metal bots. Not
sure what the direct cause is cause it was different tests that were
failing and I would have thought we'd eventually submit at the end of
ProgramsTest when tearing things down. Maybe we just created such a
backlog of work submitted to the actual GPU that the GPU got put into
a bad state?

Bug: skia:10118
Change-Id: I318428e228e08fbcaaa12b23ad2869ecc35cc786