Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > (1 comment)
> > 
> > Yeah this change is still good. I don't think we thought much about the no backbuffer case that shouldn't be happening. I feel like not having a backbuffer should be much more of an error than it currently is with just a printf? Like this is our own internal tool we should hopefully not be trying to do things that aren't support by the GPU right?
> 
> Agreed, there's a deeper bug here -- if we're given invalid settings on start up we should report that and shut down. And the GUI should only display settings that can be valid.

It appears as if changing the MSAA setting on OS X always triggers this bug. When I ran into this last night, I thought it was because I was on a Mac mini (with a puny Intel 630 GPU). I retested on my wife's Macbook Pro this morning and got the same outcome. I'll cut a bug for this.

Patch-set: 5