Update patch set 1
Patch Set 1: Commit-Queue+2

> Patch Set 1: Code-Review+1
> 
> Out of curiosity, what is the failure behavior now?

It ends up printing an error to the console every frame. It's unfortunate, but Skia as a whole is really lacking error reporting or logging. The GPU backend has a specific object for reporting errors related to shaders - if you run the SkSL slide on the GPU backend and add a typo, we show a dialog with the error message. However, that mechanism is tied to a GrContext, so when the CPU backend fails (which is what's happening here), we have no interface for giving that information back to the client. It's one of my pet peeves.

Patch-set: 1
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=+2