Update patch set 2
Patch Set 2:

So if I apply this and then https://skia-review.googlesource.com/c/skia/+/283536 everything seems to work (oddly, in a way I wouldn't expect, both with older and newer HarfBuzz) but doesn't seem to pass the layout tests. I'll take a look later to see what the differences might be, but if you can figure it out first, I'd be happy to know.

Patch-set: 2
CC: Gerrit User 5140 <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Work-in-progress: false