Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> So if I apply this and then https://skia-review.googlesource.com/c/skia/+/283536 everything seems to work (oddly, in a way I wouldn't expect, both with older and newer HarfBuzz) but doesn't seem to pass the layout tests. I'll take a look later to see what the differences might be, but if you can figure it out first, I'd be happy to know.

It might be because we're reaching the code path in Blink [1] where we trigger a makeClone() reincarnation of the font and the "trak-free metrics CTFont backed SkTypeface" does not keep this characteristic through a makeClone() cycle. 

I'll investigate when we reach this code path and whether we do a makeClone() for `opsz` even if we would not need to. 

I've also updated this sample to draw another waterfall where the SkTypeface is cycled through makeClone() in a hopefully similar fashion to what happens in Blink.


[1] https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/platform/fonts/mac/font_platform_data_mac.mm;l=177

Patch-set: 3
Reviewer: Gerrit User 6155 <6155@5086850b-e599-37d2-8ec8-ff16aeccbf02>