Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> ... are we ready to explore keeping the m44 at the device level?

We'd need to decide how to deal with type flags and queries... We lean on that heavily in the GPU backend. Options:

1) Convert to SkMatrix late, when we need to know those things.
2) Add queries to SkM44 and either:
  a) Compute the answer on the fly
  b) Add type flags/tracking to SkM44

I don't like 2b much, but I assume we have that behavior on SkMatrix for performance reasons, so 2a may not be acceptable.

Patch-set: 10
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0