Update patch set 11
Patch Set 11:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > > Patch Set 10:
> > > 
> > > I really like it. The helpers are nice. Curious what you think about passing as const& rather than const*?
> > 
> > I tried switching to const& - it may not be worth it. Some of the helpers that optionally override the provider need some extra gymnastics. Worse, the one spot where I'm already using the provider to make an Op (vertices) fails completely, because we use the templated mesh draw op helper. When doing template matching, references get converted to the underlying value type, which makes them unusable with abstract types: https://godbolt.org/z/BfP6nH
> > 
> > Maybe there's some trick to make this work? I could degrade to a pointer for this one API barrier, but if we end up using the provider in all the other RTC functions and op factories, we'd hit the same problem in many more spots.
> 
> yuck, let's abort that idea.

Ben & Mike pointed out the easy fix for this problem, so PS11 switches to const&. I think it's an improvement overall. (The refs can't be optionally re-seated to the helpers, but I made the image quad draw one unconditional, and the Android one uses a pointer temporarily. I think the clarity that this thing is never null elsewhere is worth it).

Patch-set: 11
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>