Update patch set 4
Patch Set 4: Commit-Queue+1

(1 comment)

PTAL - I significantly overhauled the unit tests. The set of compared round rects is the same, but now the verification ensures that intersect(a, b) == intersect(b, a), and result=intersect(a,b) satisfies result==intersect(result, a) and result==intersect(result, b). Additionally, uses path ops to evaluate intersect(a,b) and ensure that the conservative round rect matches.

This extra testing showed me a gap in my conservative check where if the two round rects had the same corner location and corner radii, we'd reject it because the curves couldn't be compared and the extreme corner points weren't contained in each others curves. That case is a trivial reduction of the otherwise hard generic ellipse to ellipse contains check, so I added that (it was necessary to satisfy result == intersection(result, a), etc).

To add this simple curve check, the implementation was rewritten slightly, but I think it makes it cleaner.

Patch-set: 4
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=+1