new JIT flow WIP
Main ideas:

  A) Register / stack bookkeeping separated again from
     the code generation.  Each instruction can ask for
     registers, and we'll take care of tracking where
     the values should live uniformly.

  B) No registers we use during an instruction overlap.
     This makes reasoning about correctness much easier,
     and eliminates the need for most temporaries and
     strange manual register allocation (e.g. vgatherdps).

  C) r(Val) always gives a value in a register,
     but also any(Val) lets us use it from the stack.

Redid all the register management this time around
to be as simple as possible, hopefully eliminating
bugs.  A single `regs` array tracks the state of each
register, either holding a value ID or one of a few
sentinels.

No diffs, tests pass, and perf and codegen look ok
given that hoisting is disabled.

Key lesson: hoisting is complicated!

Change-Id: Ib18c0adb1fe637a1242df3de30fbc37ccbc3c009