eliminate used_in_loop bit
We only need this used_in_loop bit now to decide whether to clean up
values in registers when they die.  Instead we can just extend the
lifetime of these values (anything that is can_hoist and used_in_loop
today) to the end of the program.

Change-Id: Ia0d29eccf81b97927add26f2f0f8226141485b8f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/284833
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Mike Klein <mtklein@google.com>