Update patch set 4
Patch Set 4:

I'm fine with this, and if existing callers are cool (or just don't care), let's rebase and land.

I do see the value in the previous behavior... you want to shove in 0s or some other value "in the native colorspace of the pixmap/bitmap'. If no one wants this, fine. If they do, we might expose a way to support that...

In that vein, what if the pixmap/bitmap is premul, and the caller wants to pass in a premul color...?

Patch-set: 4
Reviewer: Gerrit User 5649 <5649@5086850b-e599-37d2-8ec8-ff16aeccbf02>