fix GCC builds?
Looks like this is a libstdc++ bug, not being able to make a const
structured binding to OptimizedInstruction?  There are non-const
OptimizedInstruction structured bindings in the file already...

Anyway, this structured binding was already a bit of a pain for
not being able to close over its elements in a lambda.  So toss
the whole thing and manually unpack a bunch of const variables.
Kind of nice to not have to unpack fields we don't use too.

All the const here is not particularly important, but I found it
reassuring during a dark time of unknown bugs.

Change-Id: I0942ff5a0ea20f29c54d9a6a49883d17e9b40e58