Update patch set 6
Patch Set 6:

Yeah, the way I'm thinking of it is,
   - most values are dead after their last use;
   - hoisted values that are used in the loop are the exception.

Varying values are only ever valid during a single round of the loop.
Uniform and constant values (i.e. hoisted values) are valid every round of the loop.

So this CL is actually correcting the 'death' field to contain each value's death,
where previously that field contained the ID of the last instruction that uses it,
usually but not exactly its death.

Patch-set: 6
Reviewer: Gerrit User 5045 <5045@5086850b-e599-37d2-8ec8-ff16aeccbf02>